requirements.txt
absl-py==2.3.1
altair==5.3.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
filelock==3.16.1
flatbuffers==25.9.23
fonttools==4.57.0
fsspec==2025.3.0
gast==0.4.0
gitdb==4.0.12
gitpython==3.1.45
google-auth==2.43.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
idna==3.11
importlib-metadata==8.5.0
importlib-resources==6.4.5
jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==2.13.1
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.3.1
opencv-python==4.12.0.88
opt-einsum==3.4.0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pkgutil-resolve-name==1.3.10
polars==1.8.2
protobuf==4.25.8
psutil==7.1.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydeck==0.9.1
pygments==2.19.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.35.1
requests==2.32.4
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.20.1
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
smmap==5.0.2
streamlit==1.40.1
streamlit-lottie==0.0.5
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io==0.31.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==1.0.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4.2
tqdm==4.67.1
typing-extensions==4.5.0
tzdata==2025.2
ultralytics==8.3.233
ultralytics-thop==2.0.18
urllib3==2.2.3
watchdog==4.0.2
werkzeug==3.0.6
wrapt==2.0.1
zipp==3.20.2
